+++ lib/mason.dart
@@ -2,11 +2,11 @@
 /// generate files quickly and consistently.
 ///
 /// ```sh
-/// # activate mason
-/// pub global activate mason
+/// # Activate mason
+/// $ dart pub global activate mason
 ///
-/// # see usage
-/// mason --help
+/// # See usage
+/// $ mason --help
 /// ```
 library mason;