+++ test/commands/cache_test.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:io/io.dart';
 import 'package:mason/mason.dart';
+import 'package:mason/src/bricks_json.dart';
 import 'package:mason/src/command_runner.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:path/path.dart' as path;
@@ -35,7 +36,7 @@ void main() {
       Directory.current = cwd;
     });
 
-    test('clear removes .mason/brick.json', () async {
+    test('clear removes local .mason/bricks.json', () async {
       final expectedBrickJsonPath = path.join(
         Directory.current.path,
         '.mason',
@@ -51,31 +52,49 @@ void main() {
       expect(File(expectedBrickJsonPath).existsSync(), isFalse);
     });
 
-    test('clear --force removes .mason/brick.json and warns user', () async {
+    test('clear removes global .mason/bricks.json', () async {
+      final expectedBrickJsonFile = File(
+        path.join(
+          BricksJson.globalDir.path,
+          '.mason',
+          'bricks.json',
+        ),
+      );
+
+      if (!expectedBrickJsonFile.existsSync()) {
+        expectedBrickJsonFile.createSync(recursive: true);
+      }
+
+      expect(expectedBrickJsonFile.existsSync(), isTrue);
+
+      final result = await commandRunner.run(['cache', 'clear']);
+      expect(result, equals(ExitCode.success.code));
+      expect(expectedBrickJsonFile.existsSync(), isFalse);
+    });
+
+    test('clear removes .mason/bricks.json', () async {
       final expectedBrickJsonPath = path.join(
         Directory.current.path,
         '.mason',
         'bricks.json',
       );
+      if (!BricksJson.globalDir.existsSync()) {
+        BricksJson.globalDir.createSync(recursive: true);
+      }
 
       final getResult = await commandRunner.run(['get']);
       expect(getResult, equals(ExitCode.success.code));
       expect(File(expectedBrickJsonPath).existsSync(), isTrue);
+      expect(BricksJson.globalDir.existsSync(), isTrue);
 
-      final cacheClearResult = await commandRunner.run(
-        ['cache', 'clear', '--force'],
-      );
+      final cacheClearResult = await commandRunner.run(['cache', 'clear']);
       expect(cacheClearResult, equals(ExitCode.success.code));
       expect(File(expectedBrickJsonPath).existsSync(), isFalse);
       expect(
         File(path.join(Directory.current.path, 'mason.yaml')).existsSync(),
         isTrue,
       );
-      verify(
-        () => logger.warn(
-          'using --force\nI sure hope you know what you are doing.',
-        ),
-      ).called(1);
+      expect(BricksJson.globalDir.existsSync(), isFalse);
     });
   });
 }