+++ test/commands/get_test.dart
@@ -3,8 +3,9 @@ import 'dart:io';
 
 import 'package:io/io.dart';
 import 'package:mason/mason.dart';
+import 'package:mason/src/command.dart';
 import 'package:mason/src/command_runner.dart';
-import 'package:mason/src/mason_cache.dart';
+import 'package:mason/src/bricks_json.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
@@ -81,7 +82,7 @@ void main() {
       final widgetPath =
           '''widget_536b4405bffd371ab46f0948d0a5b9a2ac2cddb270ebc3d6f684217f7741422f''';
       final masonUrl = path.join(
-        MasonCache.rootDir.path,
+        BricksJson.rootDir.path,
         'git',
         '''mason_60e936dbe81fab0463b4efd5a396c50e4fcf52484fe2aa189d46874215a10b52''',
       );
@@ -143,9 +144,7 @@ void main() {
       final result = await commandRunner.run(['get']);
       expect(result, equals(ExitCode.usage.code));
       verify(
-        () => logger.err(
-          'Could not find mason.yaml.\nDid you forget to run mason init?',
-        ),
+        () => logger.err(const MasonYamlNotFoundException().message),
       ).called(1);
     });
   });